refactor(memory): remove consolidation and recall
The nightly consolidation sweep and the private-conversation recall
subrun were both invisible to the developer and neither could produce
anything for weeks. The notebook, its search index, the two memory
tools, the pre-compaction flush, and forgetting all stand.

Gone: the consolidation cluster in `packages/memory` (candidates, deep
ranking, the dream diary, promotion markers, the plan, recall signals,
ingestion, redaction, and the sweep), `recall.ts`, the brain's recall
subrun and its input marker, the store's candidate, ingestion,
tombstone, and rewrite tables with their protocol methods and client
calls, and the cron job that ran the sweep. A schema migration to
version 10 drops the tables; the retired cron row is removed at launch
so it cannot fall through to the heartbeat.

Folded: `tokenize.ts` and `vectors.ts` into `ranking.ts`, `watch.ts`
into `notebook-memory.ts`, the eligibility predicates into
`eligibility.ts`, the flush marker rows into `memory-flush-table.ts`,
and the narrowed forget into `notebook-table.ts`. The embedding
provider selection and the `unavailable` retrieval mode go with it:
`wireMemory` never passed a selection, so an adapter that cannot answer
degrades to keyword-only and says so.
