+++ packages/brain/src/agent.ts
@@ -66,7 +66,6 @@ import {
   heartbeatInputText,
   holdReleasedInputText,
   primedNotesInputText,
-  recallInputText,
   standingContextText,
   subagentTaskInputText,
   wakeInputText,
@@ -124,7 +123,6 @@ import {
   type TranscriptDeltasAttached,
 } from "./transcript-reads.js";
 import {
-  type AskTurnPlan,
   BRAIN_TURN_KIND,
   BRAIN_TURN_TRIGGER,
   type BrainTurnDescription,
@@ -339,12 +337,6 @@ export interface BrainAgentOptions {
   /** The notebook's search and read, supplied by the host that owns the index; absent, the memory tools refuse. */
   memory?: BrainMemoryAccess;
   /**
-   * Private-conversation recall for a developer's ask: the host decides
-   * whether this conversation recalls at all and runs the bounded recall,
-   * answering the summary the turn reads for one inference and never keeps.
-   */
-  recall?: (ask: BrainRecallAsk) => Promise<string | undefined>;
-  /**
    * The requester's active context a forked child starts over, adopted
    * whole into this conversation's empty context on its first turn and
    * recorded as a fork boundary; never read once the context holds anything.
@@ -410,13 +402,6 @@ function newRunControl(runId: string, generation: Generation, recorded: boolean)
     performedActs: 0,
     unknownActs: 0,
   };
-}
-
-/** What a recall is asked over: the developer's words and the run they open, for the host's cache and bounds. */
-export interface BrainRecallAsk {
-  readonly query: string;
-  readonly runId: string;
-  readonly signal: AbortSignal;
 }
 
 /** Whether a completion reached this conversation, and how. */
@@ -2075,7 +2060,6 @@ export class BrainAgent {
             : {
                 ...opened,
                 trigger: BRAIN_TURN_TRIGGER.ASK,
-                question,
                 open: (attached, now) => [askInputText(question, attached, now)],
               },
           riders,
@@ -2360,18 +2344,11 @@ export class BrainAgent {
           contextMark = context.mark();
           const primed = await this.#primeIfFresh(turnContext);
           notes = this.#options.openingNotes?.take() ?? [];
-          // Awaited only where a recall stands: an extra tick before every
-          // turn would reorder the inferences the routing tests count.
-          const recalled =
-            plan.trigger === BRAIN_TURN_TRIGGER.ASK
-              ? await this.#recallFor(plan, turnContext)
-              : undefined;
           const end = await this.#execute(turnContext, execution, gathering, {
             prompt: preparation.prompt,
             policy,
             plan,
             riders,
-            ...(recalled ? { recalled } : undefined),
             opening: [
               ...primed,
               ...(notes.length > 0 ? [activityNoticesInputText(notes, startedAt)] : []),
@@ -2544,27 +2521,6 @@ export class BrainAgent {
   }
 
   /**
-   * The recall a developer's ask earns before its inference, when the host
-   * runs one for this conversation: a summary for this turn's ephemeral
-   * context, never retained, so nothing recalled can be recalled again as if
-   * it had been said. A recall that fails, times out, or is revoked is no
-   * summary; the turn goes on without it.
-   */
-  async #recallFor(plan: AskTurnPlan, turnContext: TurnContext): Promise<string | undefined> {
-    const recall = this.#options.recall;
-    if (!recall) return undefined;
-    const settled = await settledUnlessAborted(
-      recall({
-        query: plan.question,
-        runId: turnContext.run.runId,
-        signal: turnContext.signal,
-      }).catch(() => undefined),
-      turnContext.signal,
-    );
-    return settled.aborted ? undefined : settled.value || undefined;
-  }
-
-  /**
    * One execution on the runtime: the opening words, the toolset the
    * effective policy fixes, the standing context rebuilt for every
    * inference, and a listener that keeps what the run gathers and
@@ -2580,8 +2536,6 @@ export class BrainAgent {
       plan: TurnPlan;
       riders: RunControl[];
       opening: readonly string[];
-      /** A recall's summary, rebuilt into the ephemeral context of every inference of this turn and kept nowhere. */
-      recalled?: string;
       advanceMark: () => Promise<void>;
     },
   ): Promise<RuntimeRunEnd> {
@@ -2658,7 +2612,6 @@ export class BrainAgent {
           this.#options.standingContext(),
           this.#now(),
         ),
-        ...(turn.recalled ? [recallInputText(turn.recalled, this.#now())] : []),
       ],
       maximumOutputTokens: this.#maximumOutputTokens,
       ...(this.#options.reasoningEffort
