+++ packages/memory/src/defaults.ts
@@ -22,43 +22,19 @@ export const MEMORY_SEARCH_DEFAULTS = {
 } as const;
 
 /**
- * The private-conversation recall defaults, pinned to the same revision's
- * `extensions/active-memory/types.ts`.
+ * The most characters one memory query carries, pinned to the same
+ * revision's `extensions/active-memory/types.ts`.
  */
-export const RECALL_DEFAULTS = {
-  TIMEOUT_MS: 15_000,
-  MAXIMUM_SUMMARY_CHARS: 220,
-  RECENT_USER_TURNS: 2,
-  RECENT_ASSISTANT_TURNS: 1,
-  RECENT_USER_CHARS: 220,
-  RECENT_ASSISTANT_CHARS: 180,
-  CACHE_TTL_MS: 15_000,
-  MAXIMUM_CACHE_ENTRIES: 1_000,
-  CIRCUIT_BREAKER_MAXIMUM_TIMEOUTS: 3,
-  CIRCUIT_BREAKER_COOLDOWN_MS: 60_000,
-  MAXIMUM_QUERY_CHARS: 480,
-} as const;
-
-/** How embeddings are chosen: automatically, by name, or not at all. */
-export const EMBEDDING_PROVIDER_SELECTION = {
-  AUTO: "auto",
-  OPENAI: "openai",
-  NONE: "none",
-} as const;
-
-export type EmbeddingProviderSelection =
-  (typeof EMBEDDING_PROVIDER_SELECTION)[keyof typeof EMBEDDING_PROVIDER_SELECTION];
+export const MEMORY_QUERY_MAXIMUM_CHARS = 480;
 
 /**
- * The retrieval mode a search actually ran in. Hybrid is the design; keyword
- * alone is what an automatic provider degrades to when its embeddings cannot
- * be had; unavailable is what an explicitly selected provider's failure
- * earns, because the developer asked for that provider and nothing else.
+ * The retrieval mode a search actually ran in. Hybrid is the design;
+ * keyword alone is what a search degrades to when its embeddings cannot be
+ * had, and the answer says so.
  */
 export const RETRIEVAL_MODE = {
   HYBRID: "hybrid",
   KEYWORD_ONLY: "keyword-only",
-  UNAVAILABLE: "unavailable",
 } as const;
 
 export type RetrievalMode = (typeof RETRIEVAL_MODE)[keyof typeof RETRIEVAL_MODE];
