+++ packages/memory/src/index.ts
@@ -6,66 +6,6 @@ export {
   type MemoryChunk,
 } from "./chunking.js";
 export {
-  appendOnlyPromotion,
-  applyConsolidationPlan,
-  boundCandidateText,
-  CANDIDATE_ORIGIN,
-  CANDIDATE_SESSION_KIND,
-  CANDIDATE_STATUS,
-  type CandidateOrigin,
-  type CandidateSeed,
-  type CandidateSessionKind,
-  type CandidateStatus,
-  CONSOLIDATION_ACTION,
-  CONSOLIDATION_DEFAULTS,
-  CONSOLIDATION_PHASE,
-  CONSOLIDATION_SYSTEM_PROMPT,
-  CONVERSATION_PATH_PREFIX,
-  type ConsolidationAction,
-  type ConsolidationOperation,
-  type ConsolidationPhase,
-  type ConsolidationPlan,
-  type ConsolidationResult,
-  candidateFromSeed,
-  candidateKeyFor,
-  candidateSourceRef,
-  candidatesDuplicate,
-  conceptTags,
-  consolidationJob,
-  consolidationPrompt,
-  conversationCandidatePath,
-  DEEP_RANKING_WEIGHTS,
-  type DeepRanking,
-  DREAM_DIARY_SYSTEM_PROMPT,
-  DREAMS_FILE,
-  type DreamDiaryEntry,
-  dreamDiaryEntry,
-  ingestionQuery,
-  isConsolidationCandidateEligible,
-  isConversationCandidate,
-  isPromotionOriginBlocked,
-  lineageMarker,
-  type MemoryCandidate,
-  memoryCandidateFromWire,
-  memoryEntries,
-  parseConsolidationPlan,
-  passesDeepGates,
-  promotedCandidateKeys,
-  promotedEntry,
-  promotionMarker,
-  promotionMarkerKey,
-  type RankedCandidate,
-  type RecallSignalResult,
-  type RemReflection,
-  rankCandidate,
-  recallSignalSeeds,
-  reinforceCandidate,
-  removePromotedEntries,
-  remReflections,
-  selectDeepPromotions,
-  validateConsolidationPlan,
-} from "./consolidation.js";
-export {
   type ConversationLineHit,
   type EmbeddingModelIdentity,
   type EmbeddingWrite,
@@ -89,13 +29,16 @@ export {
   type VectorHit,
 } from "./contracts.js";
 export {
-  EMBEDDING_PROVIDER_SELECTION,
-  type EmbeddingProviderSelection,
+  MEMORY_QUERY_MAXIMUM_CHARS,
   MEMORY_SEARCH_DEFAULTS,
-  RECALL_DEFAULTS,
   RETRIEVAL_MODE,
   type RetrievalMode,
 } from "./defaults.js";
+export {
+  isMaintenanceEligibleConversation,
+  isRecallEligibleConversation,
+  type RecallEligibilityInput,
+} from "./eligibility.js";
 export {
   alreadyFlushedForCompaction,
   dailyNotePathFor,
@@ -120,21 +63,6 @@ export {
   shouldRunMemoryFlush,
 } from "./flush.js";
 export {
-  boundNoteLine,
-  type DedupedSeeds,
-  dedupe,
-  historyLineHash,
-  INGESTION_SCORE,
-  type IngestibleHistoryLine,
-  type IngestibleLine,
-  ingestibleLines,
-  NOTE_LINE_MAX_CHARS,
-  type NoteFile,
-  noteSeeds,
-  recentDailyNoteDay,
-  recentHistoryLines,
-} from "./ingestion.js";
-export {
   appendNotebookEntry,
   isNotebookRootFile,
   maximumNotebookEntryLength,
@@ -151,74 +79,31 @@ export {
 export {
   CONVERSATION_RESULT_PATH_PREFIX,
   conversationResultPath,
-  embeddingUnavailable,
   type MemorySyncReport,
+  type MemoryWatcher,
+  type MemoryWatchOptions,
   NotebookMemory,
   type NotebookMemoryAccess,
   type NotebookMemoryOptions,
   type NotebookMemoryStore,
-  type NotebookRecallAsk,
-  type NotebookRecallSubrunAsk,
   type RetrievalStanding,
+  watchMemoryFiles,
 } from "./notebook-memory.js";
 export {
   bm25RankToScore,
   buildFtsQuery,
+  cosineSimilarity,
   datedNoteDay,
   decayedScore,
   defaultRankingOptions,
   type HybridRankingOptions,
   isEvergreenMemoryPath,
+  jaccardSimilarity,
   mergeHybridResults,
   mmrRerank,
+  parseEmbedding,
   selectHybridSearchResults,
+  serializeEmbedding,
+  textSimilarity,
+  tokenize,
 } from "./ranking.js";
-export {
-  boundRecentTurns,
-  ConversationRecall,
-  type ConversationRecallOptions,
-  conversationRunsRecall,
-  decideRecall,
-  hasRecallIntent,
-  isMaintenanceEligibleConversation,
-  isRecallEligibleConversation,
-  RECALL_DECISION,
-  RECALL_ELIGIBLE_KINDS,
-  RECALL_STATUS,
-  RECALL_TURN_ROLE,
-  type RecallAsk,
-  type RecallDecision,
-  type RecallEligibilityInput,
-  type RecallRecentTurn,
-  type RecallResult,
-  type RecallStatus,
-  type RecallSubrun,
-  type RecallTurnRole,
-  resolveRecallEscalation,
-  summarizeRecallReply,
-  type TrustedMemoryLookup,
-} from "./recall.js";
-export {
-  prepareForIngestion,
-  privateKeyLines,
-  RECALLED_CONTEXT_MARKER,
-  REDACTED_TOKEN,
-  type RedactionResult,
-  redactSensitiveText,
-  stripRecalledContext,
-} from "./redaction.js";
-export {
-  type ConsolidationSweepDependencies,
-  type ConsolidationSweepReport,
-  type ConsolidationSweepStore,
-  DEEP_PATH,
-  type DeepPath,
-  type DeepRewriteDecision,
-  decideDeepRewrite,
-  NOTE_BUDGET_SHARE,
-  runConsolidationSweep,
-  type ToolFreeAsk,
-} from "./sweep.js";
-export { jaccardSimilarity, stripBullet, textSimilarity, tokenize } from "./tokenize.js";
-export { cosineSimilarity, parseEmbedding, serializeEmbedding } from "./vectors.js";
-export { type MemoryWatcher, type MemoryWatchOptions, watchMemoryFiles } from "./watch.js";
