+++ packages/memory/src/notebook-memory.ts
@@ -1,3 +1,4 @@
+import { type FSWatcher, watch } from "node:fs";
 import { CONVERSATION_ENTRY_KIND, type ConversationEntry } from "@sidecar/realtime";
 import {
   type ConversationRecord,
@@ -22,41 +23,24 @@ import {
   type MemorySearchResult,
   type MemorySyncApply,
 } from "./contracts.js";
-import {
-  EMBEDDING_PROVIDER_SELECTION,
-  type EmbeddingProviderSelection,
-  MEMORY_SEARCH_DEFAULTS,
-  RETRIEVAL_MODE,
-  type RetrievalMode,
-} from "./defaults.js";
-import { selectHybridSearchResults } from "./ranking.js";
-import {
-  ConversationRecall,
-  conversationRunsRecall,
-  isRecallEligibleConversation,
-  RECALL_TURN_ROLE,
-  type RecallRecentTurn,
-} from "./recall.js";
-import { tokenize } from "./tokenize.js";
-import { type MemoryWatcher, watchMemoryFiles } from "./watch.js";
+import { MEMORY_SEARCH_DEFAULTS, RETRIEVAL_MODE, type RetrievalMode } from "./defaults.js";
+import { isRecallEligibleConversation } from "./eligibility.js";
+import { selectHybridSearchResults, tokenize } from "./ranking.js";
 
 /**
- * The notebook's index and recall as one host over injected seams. The
- * files under the agent's workspace are the source of truth; the store's
- * worker keeps the derived index and does the ranking; this class drives the
- * sync (a plan from the store, the missing vectors from the embedding
- * adapter, the apply back to the store), watches the files for a hand edit,
- * answers the brain's memory tools, and runs the private-conversation recall
- * an eligible conversation's ask earns. It knows no database, no runtime,
- * and no window: the store, the adapter, and the subrun are handed in.
+ * The notebook's index as one host over injected seams. The files under the
+ * agent's workspace are the source of truth; the store's worker keeps the
+ * derived index and does the ranking; this class drives the sync (a plan
+ * from the store, the missing vectors from the embedding adapter, the apply
+ * back to the store), watches the files for a hand edit, and answers the
+ * brain's memory tools. It knows no database, no runtime, and no window: the
+ * store and the adapter are handed in.
  *
- * Embeddings follow OpenClaw's distinction: under the automatic selection an
- * adapter that cannot answer degrades the search to keyword-only and the
- * answer says so; under an explicit selection the same failure leaves the
- * search unavailable, because the developer asked for that provider and no
- * other. No embedding is ever made of a conversation: the past-conversation
- * results a search may carry are lines already retained in History, read
- * from the store for the eligible conversations alone and indexed nowhere.
+ * An adapter that cannot answer degrades the search to keyword-only and the
+ * answer says so. No embedding is ever made of a conversation: the
+ * past-conversation results a search may carry are lines already retained in
+ * History, read from the store for the eligible conversations alone and
+ * indexed nowhere.
  */
 
 /** The store as the host reads and writes it: the index's plan and apply, its search and read, and History's search. */
@@ -90,25 +74,10 @@ export interface NotebookMemoryAccess {
   }): Promise<WireRecord>;
 }
 
-/** What the host hands the subrun it does not run itself: the asking conversation's tools, the question, and the bounds. */
-export interface NotebookRecallSubrunAsk {
-  readonly sessionKey: SessionKey;
-  readonly memory: NotebookMemoryAccess;
-  readonly query: string;
-  readonly recentTurns: readonly RecallRecentTurn[];
-  readonly signal: AbortSignal;
-}
-
-export interface NotebookRecallAsk {
-  readonly query: string;
-  readonly signal: AbortSignal;
-}
-
 export interface NotebookMemoryOptions {
   readonly store: () => NotebookMemoryStore;
   /** The embedding adapter the credential policy built, or nothing when no credential stands. */
   readonly embeddingAdapter: () => EmbeddingAdapter | undefined;
-  readonly embeddingSelection?: EmbeddingProviderSelection;
   /** The most texts one embed call carries; longer plans are cut into batches of this size. */
   readonly embeddingBatchSize: number;
   /** The agent's identity workspace, watched for the notebook's files. */
@@ -116,23 +85,10 @@ export interface NotebookMemoryOptions {
   readonly agentId?: string;
   readonly conversationDirectory: () => readonly ConversationRecord[];
   readonly isTemporary: (sessionKey: SessionKey) => boolean;
-  /** One conversation's retained lines, for the recall's small recent-turn input. */
-  readonly historyLines: (sessionKey: SessionKey) => readonly ConversationEntry[];
-  /** Runs the bounded recall subrun, or answers nothing when no runtime stands. */
-  readonly runSubrun: (ask: NotebookRecallSubrunAsk) => Promise<string | undefined>;
   readonly now: () => number;
   readonly report: (message: string) => void;
   /** Hears every completed sync, so the notebook's cached entries can be read again after a hand edit. */
   readonly onSynced?: () => void;
-  /**
-   * Hears the notebook results every search surfaced, with the query that
-   * found them, after the answer is composed: the consolidation's recall
-   * signal. Nothing it does changes what the search answered.
-   */
-  readonly onNotebookResults?: (
-    query: string,
-    results: readonly MemorySearchResult[],
-  ) => Promise<void>;
 }
 
 export interface MemorySyncReport extends MemoryApplyReport {
@@ -154,19 +110,9 @@ interface QueryEmbedding extends RetrievalStanding {
 
 const HYBRID: RetrievalStanding = { mode: RETRIEVAL_MODE.HYBRID };
 
-/** Whether embeddings are wanted at all, and what to say when they cannot be had. */
-export function embeddingUnavailable(
-  selection: EmbeddingProviderSelection,
-  reason: string,
-): RetrievalStanding {
-  if (selection === EMBEDDING_PROVIDER_SELECTION.NONE) return { mode: RETRIEVAL_MODE.KEYWORD_ONLY };
-  if (selection === EMBEDDING_PROVIDER_SELECTION.AUTO) {
-    return { mode: RETRIEVAL_MODE.KEYWORD_ONLY, note: `keyword-only: ${reason}` };
-  }
-  return {
-    mode: RETRIEVAL_MODE.UNAVAILABLE,
-    note: `the selected embedding provider failed: ${reason}`,
-  };
+/** What a search says when its embeddings cannot be had: the keyword half alone, and why. */
+function keywordOnly(reason: string): RetrievalStanding {
+  return { mode: RETRIEVAL_MODE.KEYWORD_ONLY, note: `keyword-only: ${reason}` };
 }
 
 const NO_CREDENTIAL = "no embedding credential stands";
@@ -178,8 +124,6 @@ export const CONVERSATION_RESULT_PATH_PREFIX = "conversation:";
 export function conversationResultPath(sessionKey: SessionKey): string {
   return `${CONVERSATION_RESULT_PATH_PREFIX}${sessionKey}`;
 }
-
-const RECENT_TURNS_READ = 6;
 
 /** A line's keyword score: the share of the query's tokens it carries; a search has no bm25 over History. */
 function lexicalScore(query: string, words: string): number {
@@ -255,9 +199,7 @@ function withNote(standing: RetrievalStanding): { note?: string } {
 
 export class NotebookMemory {
   readonly #options: NotebookMemoryOptions;
-  readonly #selection: EmbeddingProviderSelection;
   readonly #agentId: string;
-  readonly #recalls = new Map<SessionKey, ConversationRecall>();
   #standing: RetrievalStanding = { mode: RETRIEVAL_MODE.KEYWORD_ONLY };
   #watcher: MemoryWatcher | undefined;
   #syncing: Promise<MemorySyncReport | undefined> | undefined;
@@ -265,7 +207,6 @@ export class NotebookMemory {
 
   constructor(options: NotebookMemoryOptions) {
     this.#options = options;
-    this.#selection = options.embeddingSelection ?? EMBEDDING_PROVIDER_SELECTION.AUTO;
     this.#agentId = options.agentId ?? DEFAULT_AGENT_ID;
   }
 
@@ -369,60 +310,6 @@ export class NotebookMemory {
   }
 
   /**
-   * Forgets every cached recall, after a forget or a durable rewrite changed
-   * what a recall would say. The recalls stay where the conversations that
-   * bound them can reach them; each drops its cached answers and refuses to
-   * cache a run that began before the clear, so a recall still in flight
-   * settles for its own callers and leaves nothing stale behind.
-   */
-  clearRecallCaches(): void {
-    for (const recall of this.#recalls.values()) recall.invalidate();
-  }
-
-  /** The recall an eligible conversation's asks run, or nothing for one that does not recall. */
-  recallFor(
-    sessionKey: SessionKey,
-  ): ((ask: NotebookRecallAsk) => Promise<string | undefined>) | undefined {
-    if (
-      !conversationRunsRecall({
-        sessionKey,
-        agentId: this.#agentId,
-        temporary: this.#options.isTemporary(sessionKey),
-      })
-    ) {
-      return undefined;
-    }
-    let recall = this.#recalls.get(sessionKey);
-    if (!recall) {
-      recall = new ConversationRecall({
-        now: this.#options.now,
-        report: this.#options.report,
-        trustedMemory: (query, signal) => this.#trustedMemory(query, signal),
-        subrun: ({ query, recentTurns, signal }) =>
-          this.#options.runSubrun({
-            sessionKey,
-            memory: this.accessFor(sessionKey),
-            query,
-            recentTurns,
-            signal,
-          }),
-      });
-      this.#recalls.set(sessionKey, recall);
-    }
-    const bound = recall;
-    return async (ask) => {
-      const result = await bound.recall({
-        sessionKey,
-        agentId: this.#agentId,
-        query: ask.query,
-        recentTurns: this.#recentTurns(sessionKey),
-        signal: ask.signal,
-      });
-      return result.summary.length > 0 ? result.summary : undefined;
-    };
-  }
-
-  /**
    * One search from `current`: the notebook's chunks and the eligible
    * conversations' lines ranked inside the same window, under the same
    * weights and the same threshold. The mode is this call's own; it moves
@@ -434,9 +321,6 @@ export class NotebookMemory {
   ): Promise<MemorySearchAnswer> {
     const maxResults = ask.maxResults ?? MEMORY_SEARCH_DEFAULTS.MAXIMUM_RESULTS;
     const embedding = await this.#embedQuery(ask.query, ask.signal);
-    if (embedding.mode === RETRIEVAL_MODE.UNAVAILABLE) {
-      return { mode: embedding.mode, results: [], ...withNote(embedding) };
-    }
     const [notebook, conversations] = await Promise.all([
       this.#searchNotebook(ask.query, embedding, maxResults),
       this.#conversationResults(
@@ -448,7 +332,6 @@ export class NotebookMemory {
     const merged = [...notebook.results, ...conversations].sort(
       (a, b) => b.score - a.score || a.path.localeCompare(b.path) || a.startLine - b.startLine,
     );
-    void this.#options.onNotebookResults?.(ask.query, notebook.results);
     return {
       mode: embedding.mode,
       results: selectHybridSearchResults({
@@ -458,29 +341,6 @@ export class NotebookMemory {
         minScore: MEMORY_SEARCH_DEFAULTS.MINIMUM_SCORE,
       }),
       ...withNote(embedding),
-    };
-  }
-
-  /** The trusted lookup before any subrun: the notebook's own index, and nothing said in a conversation. */
-  async #trustedMemory(
-    query: string,
-    signal: AbortSignal | undefined,
-  ): Promise<{ strongHit: boolean }> {
-    const embedding = await this.#embedQuery(query, signal ?? new AbortController().signal);
-    if (embedding.mode === RETRIEVAL_MODE.UNAVAILABLE) return { strongHit: false };
-    const notebook = await this.#searchNotebook(
-      query,
-      embedding,
-      MEMORY_SEARCH_DEFAULTS.MAXIMUM_RESULTS,
-    );
-    // The lookup is a real retrieval from the notebook, so what it surfaced is
-    // a recall signal like any search's; the conversation fallback never runs
-    // here, so no line of History can be mistaken for one.
-    void this.#options.onNotebookResults?.(query, notebook.results);
-    return {
-      strongHit: notebook.results.some(
-        (result) => result.score >= MEMORY_SEARCH_DEFAULTS.MINIMUM_SCORE,
-      ),
     };
   }
 
@@ -501,21 +361,18 @@ export class NotebookMemory {
 
   /** The query's vector under the adapter that stands now, or the standing its absence or failure earns. */
   async #embedQuery(query: string, signal: AbortSignal): Promise<QueryEmbedding> {
-    if (this.#selection === EMBEDDING_PROVIDER_SELECTION.NONE) {
-      return embeddingUnavailable(this.#selection, NO_CREDENTIAL);
-    }
     const adapter = this.#options.embeddingAdapter();
-    if (!adapter) return embeddingUnavailable(this.#selection, NO_CREDENTIAL);
+    if (!adapter) return keywordOnly(NO_CREDENTIAL);
     const answer = await adapter.embed([query], { signal });
     if (answer.outcome === MODEL_RESPONSE_OUTCOME.THROTTLED) {
-      return embeddingUnavailable(this.#selection, RATE_LIMITED);
+      return keywordOnly(RATE_LIMITED);
     }
     if (answer.outcome === MODEL_RESPONSE_OUTCOME.FAILED) {
-      return embeddingUnavailable(this.#selection, `${answer.failure}: ${answer.reason}`);
+      return keywordOnly(`${answer.failure}: ${answer.reason}`);
     }
     const queryVector = answer.vectors[0];
     if (!queryVector) {
-      return embeddingUnavailable(this.#selection, "the embedding provider answered no vector");
+      return keywordOnly("the embedding provider answered no vector");
     }
     return { ...HYBRID, queryVector, identity: await identityOf(adapter) };
   }
@@ -526,15 +383,10 @@ export class NotebookMemory {
     // One read of the adapter for the whole pass: the identity the plan is
     // asked under and the adapter the vectors come from are the same one,
     // whatever a credential swap installs meanwhile.
-    const adapter =
-      this.#selection === EMBEDDING_PROVIDER_SELECTION.NONE
-        ? undefined
-        : this.#options.embeddingAdapter();
+    const adapter = this.#options.embeddingAdapter();
     const identity = adapter ? await identityOf(adapter) : undefined;
     const plan = await store.planMemorySync(identity, now);
-    let standing: RetrievalStanding = adapter
-      ? HYBRID
-      : embeddingUnavailable(this.#selection, NO_CREDENTIAL);
+    let standing: RetrievalStanding = adapter ? HYBRID : keywordOnly(NO_CREDENTIAL);
     let embeddings: readonly EmbeddingWrite[] = [];
     if (adapter && plan.missingEmbeddings.length > 0) {
       const embedded = await this.#embedAll(adapter, plan.missingEmbeddings);
@@ -543,7 +395,7 @@ export class NotebookMemory {
       // already cached is kept on its chunk; only the chunks still without
       // one are asked for again next sync.
       embeddings = embedded.written;
-      if (embedded.failed) standing = embeddingUnavailable(this.#selection, embedded.failed);
+      if (embedded.failed) standing = keywordOnly(embedded.failed);
     }
     const report = await store.applyMemorySync({
       changed: plan.changed,
@@ -607,23 +459,55 @@ export class NotebookMemory {
     const hits = await this.#options.store().searchHistory(keys, query, limit, this.#options.now());
     return hits.map((hit, ordinal) => conversationResult(query, hit, ordinal));
   }
-
-  #recentTurns(sessionKey: SessionKey): RecallRecentTurn[] {
-    return this.#options
-      .historyLines(sessionKey)
-      .slice(-RECENT_TURNS_READ)
-      .flatMap((entry): RecallRecentTurn[] => {
-        if (isAsk(entry)) return [{ role: RECALL_TURN_ROLE.USER, text: entry.words }];
-        if (entry.kind === CONVERSATION_ENTRY_KIND.REPLY) {
-          return [{ role: RECALL_TURN_ROLE.ASSISTANT, text: entry.words }];
-        }
-        return [];
-      });
-  }
 }
 
 /** The identity the index stores beside a vector: the adapter's provider and model, never its width. */
 async function identityOf(adapter: EmbeddingAdapter): Promise<EmbeddingModelIdentity> {
   const identity = await adapter.identity();
   return { provider: identity.provider, model: identity.model };
+}
+
+/**
+ * Watches the notebook's directory and asks for one reconcile per burst of
+ * changes, debounced at the pinned 1,500 ms. The watcher decides nothing
+ * about what changed: the reconcile it triggers reads every file again and
+ * compares hashes, so a missed event costs a later pass and never a wrong
+ * index, and an index can always be rebuilt from the files alone.
+ */
+export interface MemoryWatcher {
+  close(): void;
+}
+
+export interface MemoryWatchOptions {
+  readonly directory: string;
+  readonly onChange: () => void;
+  readonly debounceMs?: number;
+  readonly report?: (message: string) => void;
+}
+
+export function watchMemoryFiles(options: MemoryWatchOptions): MemoryWatcher | undefined {
+  const debounceMs = options.debounceMs ?? MEMORY_SEARCH_DEFAULTS.WATCH_DEBOUNCE_MS;
+  let timer: ReturnType<typeof setTimeout> | undefined;
+  let watcher: FSWatcher;
+  try {
+    watcher = watch(options.directory, { recursive: true, persistent: false }, () => {
+      if (timer) clearTimeout(timer);
+      timer = setTimeout(() => {
+        timer = undefined;
+        options.onChange();
+      }, debounceMs);
+    });
+  } catch (error) {
+    options.report?.(
+      `Memory files are not being watched: ${error instanceof Error ? error.message : String(error)}`,
+    );
+    return undefined;
+  }
+  watcher.on("error", (error) => options.report?.(`Memory watcher stopped: ${error.message}`));
+  return {
+    close: () => {
+      if (timer) clearTimeout(timer);
+      watcher.close();
+    },
+  };
 }
