+++ packages/runtime-store/src/index.ts
@@ -2,6 +2,7 @@ export type { DeletionOutcome } from "./archives.js";
 export { RuntimeStoreClient } from "./client.js";
 export type { HistorySearchHit } from "./history-table.js";
 export type { MaintenanceReport } from "./maintenance-run.js";
+export type { FlushState } from "./memory-flush-table.js";
 export type {
   EmbeddingWrite,
   MemoryApplyReport,
@@ -10,19 +11,11 @@ export type {
   MemorySearchOutcome,
   MemorySearchQuery,
 } from "./memory-index-table.js";
-export {
-  type CandidateStagingReport,
-  type FlushState,
-  FORGOTTEN_SOURCE_KIND,
-  type ForgottenSource,
-  type ForgottenSourceKind,
-  MEMORY_REWRITE_REFUSAL,
-  type MemoryForgetAsk,
-  type MemoryForgetReport,
-  type MemoryRewriteAsk,
-  type MemoryRewriteOutcome,
-  type MemoryRewriteRecord,
-} from "./memory-maintenance-table.js";
-export type { NotebookEntry, NotebookMutation } from "./notebook-table.js";
+export type {
+  MemoryForgetAsk,
+  MemoryForgetReport,
+  NotebookEntry,
+  NotebookMutation,
+} from "./notebook-table.js";
 export type { RuntimeStorePort } from "./protocol.js";
 export { serveRuntimeStore } from "./worker-host.js";
